# docs/swagger.yaml
basePath: /api
definitions:
  datatypes.JSONMap:
    additionalProperties: true
    type: object
...
  github_com_smallbiznis_railzway_internal_invoice_domain.Invoice:
    properties:
      billingCycleID:
        type: integer
      createdAt:
        type: string
      currency:
        type: string
      customerID:
        type: integer
      dueAt:
        type: string
      finalizedAt:
        type: string
      id:
        type: integer
      invoiceNumber:
        type: string
      invoiceSeq:
        type: integer
      invoiceTemplateID:
        type: integer
      issuedAt:
        type: string
      items:
        description: Items is populated for API responses, not persisted
        items:
          $ref: '#/definitions/github_com_smallbiznis_railzway_internal_invoice_domain.InvoiceItem'
        type: array
      metadata:
        $ref: '#/definitions/datatypes.JSONMap'
      orgID:
        type: integer
      paidAt:
        type: string
      periodEnd:
        type: string
      periodStart:
        type: string
      renderedHTML:
        type: string
      renderedPDFURL:
        type: string
      status:
        $ref: '#/definitions/github_com_smallbiznis_railzway_internal_invoice_domain.InvoiceStatus'
      subscriptionID:
        type: integer
      subtotalAmount:
        type: integer
      taxAmount:
        type: integer
      taxCode:
        type: string
      taxRate:
        type: number
      totalAmount:
        type: integer
      updatedAt:
        type: string
      voidedAt:
        type: string
    type: object
  github_com_smallbiznis_railzway_internal_invoice_domain.InvoiceItem:
    properties:
      amount:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      invoiceID:
        type: integer
      lineType:
        $ref: '#/definitions/github_com_smallbiznis_railzway_internal_invoice_domain.InvoiceItemLineType'
      metadata:
        $ref: '#/definitions/datatypes.JSONMap'
      orgID:
        type: integer
      quantity:
        type: number
      ratingResultID:
        type: integer
      unitPrice:
        type: integer
    type: object
  github_com_smallbiznis_railzway_internal_invoice_domain.InvoiceItemLineType:
    enum:
    - subscription
    - usage
    - credit
    - one_off
    - tax
    type: string
    x-enum-varnames:
    - InvoiceItemLineTypeSubscription
    - InvoiceItemLineTypeUsage
    - InvoiceItemLineTypeCredit
    - InvoiceItemLineTypeOneOff
    - InvoiceItemLineTypeTax
  github_com_smallbiznis_railzway_internal_invoice_domain.InvoiceStatus:
    enum:
    - DRAFT
    - FINALIZED
    - VOID
    type: string
    x-enum-varnames:
    - InvoiceStatusDraft
    - InvoiceStatusFinalized
    - InvoiceStatusVoid
  github_com_smallbiznis_railzway_internal_invoice_domain.RenderInvoiceResponse:
    properties:
      invoice_template_id:
        type: string
      is_snapshot:
        type: boolean
      rendered_html:
        type: string
      rendered_pdf_url:
        type: string
    type: object
...
  github_com_smallbiznis_railzway_internal_price_domain.CreateRequest:
    properties:
      active:
        type: boolean
      aggregate_usage:
        type: string
      billing_interval:
        type: string
      billing_interval_count:
        type: integer
      billing_mode:
        type: string
      billing_threshold:
        type: number
      billing_unit:
        type: string
      code:
        type: string
      description:
        type: string
      is_default:
        type: boolean
      lookup_key:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
      pricing_model:
        type: string
      product_id:
        type: string
      retired_at:
        type: string
      tax_behavior:
        type: string
      tax_code:
        type: string
      version:
        type: integer
    type: object
...
  github_com_smallbiznis_railzway_internal_price_domain.Response:
    properties:
      active:
        type: boolean
      aggregate_usage:
        type: string
      billing_interval:
        type: string
      billing_interval_count:
        type: integer
      billing_mode:
        type: string
      billing_threshold:
        type: number
      billing_unit:
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      lookup_key:
        type: string
      name:
        type: string
      organization_id:
        type: integer
      pricing_model:
        type: string
      product_id:
        type: integer
      retired_at:
        type: string
      tax_behavior:
        type: string
      tax_code:
        type: string
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  github_com_smallbiznis_railzway_internal_reference_domain.Country:
    properties:
      code:
        type: string
      created_at:
        type: string
      name:
        type: string
    type: object
  github_com_smallbiznis_railzway_internal_reference_domain.Currency:
    properties:
      code:
        type: string
      created_at:
        type: string
      is_active:
        type: boolean
      minor_unit:
        type: integer
      name:
        type: string
      symbol:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@railzway.com
    name: API Support
  description: Railzway Billing & Operations API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Railzway API
  version: "1.0"
paths:
  /countries:
    get:
      consumes:
      - application/json
      description: List available countries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_smallbiznis_railzway_internal_reference_domain.Country'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List Countries
      tags:
      - reference
  /currencies:
    get:
      consumes:
      - application/json
      description: List available currencies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_smallbiznis_railzway_internal_reference_domain.Currency'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List Currencies
      tags:
      - reference
...
  /invoices:
    get:
      consumes:
      - application/json
      description: List available invoices
      parameters:
      - description: Status
        in: query
        name: status
        type: string
      - description: Invoice Number
        in: query
        name: invoice_number
        type: string
      - description: Customer ID
        in: query
        name: customer_id
        type: string
      - description: Created From
        in: query
        name: created_from
        type: string
      - description: Created To
        in: query
        name: created_to
        type: string
      - description: Due From
        in: query
        name: due_from
        type: string
      - description: Due To
        in: query
        name: due_to
        type: string
      - description: Finalized From
        in: query
        name: finalized_from
        type: string
      - description: Finalized To
        in: query
        name: finalized_to
        type: string
      - description: Total Min
        in: query
        name: total_min
        type: integer
      - description: Total Max
        in: query
        name: total_max
        type: integer
      - description: Page Token
        in: query
        name: page_token
        type: string
      - description: Page Size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_smallbiznis_railzway_internal_invoice_domain.Invoice'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List Invoices
      tags:
      - invoices
  /invoices/{id}:
    get:
      consumes:
      - application/json
      description: Get invoice by ID
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_smallbiznis_railzway_internal_invoice_domain.Invoice'
      security:
      - ApiKeyAuth: []
      summary: Get Invoice
      tags:
      - invoices
  /invoices/{id}/render:
    get:
      consumes:
      - application/json
      description: Render invoice PDF/HTML
      parameters:
      - description: Invoice ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_smallbiznis_railzway_internal_invoice_domain.RenderInvoiceResponse'
      security:
      - ApiKeyAuth: []
      summary: Render Invoice
      tags:
      - invoices
...
  /pricings:
    get:
      consumes:
      - application/json
      description: List available pricings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_smallbiznis_railzway_internal_price_domain.Response'
            type: array
      security:
      - ApiKeyAuth: []
      summary: List Pricings
      tags:
      - pricings
    post:
      consumes:
      - application/json
      description: Create a new pricing
      parameters:
      - description: Create Pricing Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_smallbiznis_railzway_internal_price_domain.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_smallbiznis_railzway_internal_price_domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Pricing
      tags:
      - pricings
  /pricings/{id}:
    get:
      consumes:
      - application/json
      description: Get pricing by ID
      parameters:
      - description: Pricing ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_smallbiznis_railzway_internal_price_domain.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Pricing
      tags:
      - pricings
...
  /timezones:
    get:
      consumes:
      - application/json
      description: List timezones for a country
      parameters:
      - description: Country Code
        in: query
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - ApiKeyAuth: []
      summary: List Timezones
      tags:
      - reference
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"